Split SetupOnlyForOverridableMembers tests
