+++ Source/Moq.Analyzers.Test/SetupShouldBeUsedOnlyForOverridableMembersAnalyzerTests.ShouldFailWhenSetupIsCalledWithANonVirtualProperty.verified.txt
@@ -1,0 +1,8 @@
+﻿Diagnostic 1
+	Id: Moq1200
+	Location: SourceFile(Test0.cs[785..795))
+	Highlight: x.Property
+	Lines: mock.Setup(x => x.Property);
+	Severity: Error
+	Message: Setup should be used only for overridable members.
+
