+++ Source/Moq.Analyzers.Test/SetupShouldBeUsedOnlyForOverridableMembersAnalyzerTests.ShouldFailWhenSetupIsCalledWithASealedMethod.verified.txt
@@ -1,0 +1,8 @@
+﻿Diagnostic 1
+	Id: Moq1200
+	Location: SourceFile(Test0.cs[773..835))
+	Highlight: x.Calculate(It.IsAny<int>(), It.IsAny<int>(), It.IsAny<int>())
+	Lines: mock.Setup(x => x.Calculate(It.IsAny<int>(), It.IsAny<int>(), It.IsAny<int>()));
+	Severity: Error
+	Message: Setup should be used only for overridable members.
+
