--- Source/Moq.Analyzers.Test/SetupShouldBeUsedOnlyForOverridableMembersAnalyzerTests.ShouldPassIfGoodParameters.verified.txt
@@ -1,24 +1,0 @@
-﻿Diagnostic 1
-	Id: Moq1200
-	Location: SourceFile(Test0.cs[1571..1584))
-	Highlight: x.Calculate()
-	Lines: mock.Setup(x => x.Calculate());
-	Severity: Error
-	Message: Setup should be used only for overridable members.
-
-Diagnostic 2
-	Id: Moq1200
-	Location: SourceFile(Test0.cs[1715..1777))
-	Highlight: x.Calculate(It.IsAny<int>(), It.IsAny<int>(), It.IsAny<int>())
-	Lines: mock.Setup(x => x.Calculate(It.IsAny<int>(), It.IsAny<int>(), It.IsAny<int>()));
-	Severity: Error
-	Message: Setup should be used only for overridable members.
-
-Diagnostic 3
-	Id: Moq1200
-	Location: SourceFile(Test0.cs[1914..1924))
-	Highlight: x.Property
-	Lines: mock.Setup(x => x.Property);
-	Severity: Error
-	Message: Setup should be used only for overridable members.
-
