+++ Source/Moq.Analyzers.Test/SetupShouldBeUsedOnlyForOverridableMembersAnalyzerTests.ShouldPassWhenSetupIsCalledWithAVirtualMethod.verified.txt
@@ -1,0 +1,1 @@
+﻿emptyString
