+++ Source/Moq.Analyzers.Test/SetupShouldBeUsedOnlyForOverridableMembersAnalyzerTests.ShouldPassWhenSetupIsCalledWithAnAbstractMethod.verified.txt
@@ -1,0 +1,1 @@
+﻿emptyString
