+++ Source/Moq.Analyzers.Test/SetupShouldBeUsedOnlyForOverridableMembersAnalyzerTests.ShouldPassWhenSetupIsCalledWithAnInterfaceMethod.verified.txt
@@ -1,0 +1,1 @@
+﻿emptyString
