+++ Source/Moq.Analyzers.Test/SetupShouldBeUsedOnlyForOverridableMembersAnalyzerTests.ShouldPassWhenSetupIsCalledWithAnInterfaceProperty.verified.txt
@@ -1,0 +1,1 @@
+﻿emptyString
