+++ Source/Moq.Analyzers.Test/SetupShouldBeUsedOnlyForOverridableMembersAnalyzerTests.ShouldPassWhenSetupIsCalledWithAnOverrideOfAnAbstractMethod.verified.txt
@@ -1,0 +1,1 @@
+﻿emptyString
